First commit: Test libraries for Twitter scraping
Explore various scraping libraries, such as Ntscraper, Scrapfly, and
Snscrape, to determine which offers the best functionality and ease of
use for retrieving public Twitter data.

The goal is to compare performance, ease of authentication, and each
library's ability to integrate with the project's current architecture.
These experiments will guide the choice of the library best suited for
our data and scalability requirements.